# api/openapi.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Handy App API page
paths:
  /:
    get:
      summary: home page
      description: A home page.
      tags:
        - Home
      responses:
        "200":
          description: A home page response
  /api/user:
    post:
      summary: Create a single user
      description: Create a single user in db
      tags:
        - User
      responses:
        "201":
          description: A User has been created
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                contactNum:
                  type: string
                pwd:
                  type: string
                  format: password
                stAddress:
                  type: string
                city:
                  type: string
                postCode:
                  type: string
                bookings:
                  type: array
                  items:
                    type: string
                orders:
                  type: array
                  items:
                    type: string
              required:
                - firstName
                - lastName
                - email
                - contactNum
                - pwd
                - stAddress
                - city
                - postCode
  /api/user/orders/{userId}:
    get:
      summary: List all of a user's bookings
      description: List all of a given user id's bookings
      tags:
        - User
      responses:
        "200":
          description: These are this user's bookings
      parameters:
        - in: path
          name: userId
          description: User's ObjectId in db
          required: true
          schema:
            type: string
  /api/pro:
    post:
      summary: Create a single pro
      description: Create a single pro in db
      tags:
        - Pro
      responses:
        "201":
          description: A new pro has been created
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                contactNum:
                  type: string
                pwd:
                  type: string
                  format: password
                stAddress:
                  type: string
                city:
                  type: string
                postCode:
                  type: string
                bookings:
                  type: array
                  items:
                    type: string
                orders:
                  type: array
                  items:
                    type: string
                serviceType:
                  type: string
                availability:
                  type: array
                  items:
                    type: object
                    properties:
                      startDate:
                        type: string
                        format: date
                      startSession:
                        type: number
                      endDate:
                        type: string
                        format: date
                      endSession:
                        type: number
              required:
                - firstName
                - lastName
                - email
                - contactNum
                - pwd
                - stAddress
                - city
                - postCode
  /api/booking:
    post:
      summary: create a single booking
      description: create a single booking, for simplicity's sake, only same day, non-repetitive job is supported
      tags:
        - Booking
      responses:
        "201":
          description: The booking with our pro has been successfully created.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                customerId:
                  type: string
                proId:
                  type: string
                bookingDate:
                  type: string
                  format: date
                startSession:
                  type: number
                endSession:
                  type: number
                notes:
                  type: string
                feedback:
                  type: string
              required:
                - customerId
                - proId
                - bookingDate
                - startSession
                - endSession